Rename file library-example.model.ts -> library-item.model.ts
