+++ src/app/_common/services/library-menu.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 
-import { LibraryItem } from 'src/app/shared/models/library-example.model';
+import { LibraryItem } from 'src/app/shared/models/library-item.model';
 import { LibraryMenuLevel2Item } from 'src/app/shared/models/library-menu-level-2-item.model';
 import { LibraryMenuLevel2Category } from 'src/app/shared/models/library-menu-level-2-category.model';
 import { LibraryMenuLevel2Categories } from 'src/app/_common/constants/library-menu-level-2-categories';
