+++ src/app/shared/models/library-item.model.ts
