+++ src/mock-db-data/library-css/colors-fonts/fhi-colors.ts
@@ -1,4 +1,4 @@
-import { LibraryItem } from 'src/app/shared/models/library-example.model';
+import { LibraryItem } from 'src/app/shared/models/library-item.model';
 
 const colorHtml = (hex: string, color: string) => {
   return `
