+++ src/mock-db-data/library-css/css-components.data.ts
@@ -1,4 +1,4 @@
-import { LibraryItem } from 'src/app/shared/models/library-example.model';
+import { LibraryItem } from 'src/app/shared/models/library-item.model';
 
 import { Accordion } from './components/accordion';
 import { FhiBtnPrimary } from './components/fhi-btn-primary';
