+++ src/mock-db-data/library-css/css-icons.data.ts
@@ -1,4 +1,4 @@
-import { LibraryItem } from 'src/app/shared/models/library-example.model';
+import { LibraryItem } from 'src/app/shared/models/library-item.model';
 
 import { IconsSet } from './icons/icon-set';
