+++ src/mock-db-data/library-css/icons/icon-set.ts
@@ -1,4 +1,4 @@
-import { LibraryItem } from 'src/app/shared/models/library-example.model';
+import { LibraryItem } from 'src/app/shared/models/library-item.model';
 
 export const IconsSet: LibraryItem[] = [{
   title: 'Icon set',
