+++ src/mock-db-data/library-css/page-templates/fhi-btn-primary.ts
@@ -1,4 +1,4 @@
-import { LibraryItem } from 'src/app/shared/models/library-example.model';
+import { LibraryItem } from 'src/app/shared/models/library-item.model';
 
 export const FhiBtnPrimary: LibraryItem[] = [{
   title: 'Hovedknapp',
